Refactor add-on structure
