+++ changelog.txt
@@ -1,5 +1,6 @@
 v0.5.0
 - Update to use SPDX license identifiers
+- Refactor add-on structure
 
 v0.4.6 to v0.4.7
 - Update version list
