+++ resources/lib/version_check/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+"""
+
+    Copyright (C) 2013-2014 Team-XBMC
+    Copyright (C) 2014-2019 Team Kodi
+
+    This file is part of service.xbmc.versioncheck
+
+    SPDX-License-Identifier: GPL-3.0-or-later
+    See LICENSES/GPL-3.0-or-later.txt for more information.
+
+"""
+
+__all__ = ['aptdaemonhandler', 'common', 'jsoninterface', 'service', 'shellhandlerapt', 'versions', 'viewer']
